// src/Tools/Brush/Brush.d.ts
import type { RawVoxelData } from "../../Voxels/Types/Voxel.types.js";
import { WorldCursor } from "../../World/Cursor/WorldCursor.js";
import { VoxelCursor } from "../../Voxels/Cursor/VoxelCursor.js";
import { IVoxelTemplate } from "../../Templates/VoxelTemplates.types.js";
import { PaintVoxelData } from "../../Voxels/Types/PaintVoxelData.js";
import { VoxelPathData } from "../../Templates/Path/VoxelPath.types.js";
import { IVoxelSelection } from "../../Templates/Selection/VoxelSelection.js";
export declare class BrushTool {
    data: PaintVoxelData;
    dimension: number;
    x: number;
    y: number;
    z: number;
    voxelCursor: VoxelCursor;
    dataCursor: WorldCursor;
    setXYZ(x: number, y: number, z: number): this;
    newSector(): void;
    setData(data: Partial<PaintVoxelData>): this;
    getData(): PaintVoxelData;
    setRaw(data: RawVoxelData): this;
    getRaw(): RawVoxelData;
    setId(id: string): this;
    setName(name: string): this;
    setSecondaryId(id: string): this;
    setSecondaryName(name: string): this;
    setShapeState(state: number): this;
    setLevel(level: number): this;
    setLevelState(levelState: number): this;
    setMod(mod: number): this;
    clear(): void;
    protected _paint(): false | undefined;
    protected _paintRaw(raw: RawVoxelData): false | undefined;
    protected _erase(): void;
    paint(): this;
    paintTemplate(voxelTemplate: IVoxelTemplate): void;
    paintPath(data: VoxelPathData): void;
    erase(): this;
    eraseTemplate(voxelTemplate: IVoxelTemplate): void;
    eraseSelection(selection: IVoxelSelection): void;
    erasePath(data: VoxelPathData): void;
    start(dimension: number, x: number, y: number, z: number): this;
    stop(): this;
}